requirements.txt
aiosqlite>=0.22.1
httpx>=0.28.1
pycountry>=24.6.1
fastapi>=0.128.0
uvicorn[standard]>=0.40.0
SQLAlchemy>=2.0.46
asyncpg>=0.31.0
psycopg2-binary>=2.9.11
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.5.0

python-multipart>=0.0.22
email-validator>=2.1.1
